# packages/contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: SaaS Core API
  version: 1.0.0
  description: Wersjonowane API modularnej platformy SaaS Core.
paths:
...
  /api/v1/sites/:
    get:
      operationId: sites_list
      parameters:
      - in: query
        name: cursor
        schema:
          type: string
          format: uuid
      - in: query
        name: limit
        schema:
          type: integer
        description: Liczba elementów od 1 do 100; domyślnie 50.
      tags:
      - sites
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteList'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
    post:
      operationId: sites_create
      parameters:
      - in: header
        name: Idempotency-Key
        schema:
          type: string
        description: Klucz bezpiecznego ponowienia mutacji w zakresie organizacji
          i użytkownika.
        required: true
      tags:
      - sites
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SiteCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SiteCreate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteSummary'
          description: ''
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteSummary'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
  /api/v1/sites/{site_id}/localization/:
    get:
      operationId: sites_localization_report_retrieve
      parameters:
      - in: path
        name: site_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - sites
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteLocalizationReport'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
  /api/v1/sites/{site_id}/pages/:
    get:
      operationId: sites_pages_list
      parameters:
      - in: query
        name: cursor
        schema:
          type: string
          format: uuid
      - in: query
        name: limit
        schema:
          type: integer
        description: Liczba elementów od 1 do 100; domyślnie 50.
      - in: path
        name: site_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - sites
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageList'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
    post:
      operationId: sites_pages_create
      parameters:
      - in: header
        name: Idempotency-Key
        schema:
          type: string
        description: Klucz bezpiecznego ponowienia mutacji w zakresie organizacji
          i użytkownika.
        required: true
      - in: path
        name: site_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - sites
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PageCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PageCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PageCreate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageSummary'
          description: ''
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageSummary'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
  /api/v1/sites/pages/{page_id}/draft/:
    get:
      operationId: sites_page_draft_retrieve
      parameters:
      - in: path
        name: page_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - sites
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageDraft'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
    put:
      operationId: sites_page_draft_save
      parameters:
      - in: header
        name: Idempotency-Key
        schema:
          type: string
        description: Klucz bezpiecznego ponowienia mutacji w zakresie organizacji
          i użytkownika.
        required: true
      - in: path
        name: page_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - sites
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftSave'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DraftSave'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DraftSave'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageDraft'
          description: ''
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageDraft'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
  /api/v1/sites/pages/{page_id}/preview/{version_id}/:
    get:
      operationId: sites_page_draft_preview_retrieve
      parameters:
      - in: path
        name: page_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: version_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - sites
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageDraft'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
  /api/v1/sites/pages/{page_id}/translations/:
    get:
      operationId: sites_page_translations_list
      parameters:
      - in: path
        name: page_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - sites
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageTranslationList'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
  /api/v1/sites/pages/{page_id}/translations/{locale}/:
    put:
      operationId: sites_page_translation_save
      parameters:
      - in: header
        name: Idempotency-Key
        schema:
          type: string
        description: Klucz bezpiecznego ponowienia mutacji w zakresie organizacji
          i użytkownika.
        required: true
      - in: path
        name: locale
        schema:
          type: string
        required: true
      - in: path
        name: page_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - sites
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PageTranslationSave'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PageTranslationSave'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PageTranslationSave'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageTranslation'
          description: ''
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageTranslation'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
components:
  schemas:
...
    DraftSave:
      type: object
      properties:
        expected_version:
          type: integer
          minimum: 0
        blocks:
          type: array
          items:
            $ref: '#/components/schemas/PageBlockInput'
        media_asset_ids:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 100
      required:
      - blocks
      - expected_version
...
    LocaleEnum:
      enum:
      - pl
      - en
      type: string
      description: |-
        * `pl` - pl
        * `en` - en
...
    PageBlock:
      type: object
      properties:
        id:
          type: string
          format: uuid
        position:
          type: integer
        block_type:
          type: string
        schema_version:
          type: integer
        data: {}
      required:
      - block_type
      - data
      - id
      - position
      - schema_version
...
    PageCreate:
      type: object
      properties:
        name:
          type: string
          maxLength: 160
        key:
          type: string
          maxLength: 80
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
      required:
      - key
      - name
    PageDraft:
      type: object
      properties:
        page_id:
          type: string
          format: uuid
        version:
          type: integer
        draft_id:
          type:
          - string
          - 'null'
          format: uuid
        content_hash:
          type:
          - string
          - 'null'
        created_at:
          type:
          - string
          - 'null'
          format: date-time
        blocks:
          type: array
          items:
            $ref: '#/components/schemas/PageBlock'
        media_asset_ids:
          type: array
          items:
            type: string
            format: uuid
      required:
      - blocks
      - content_hash
      - created_at
      - draft_id
      - media_asset_ids
      - page_id
      - version
    PageList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PageSummary'
        next_cursor:
          type:
          - string
          - 'null'
          format: uuid
      required:
      - items
      - next_cursor
    PageLocalization:
      type: object
      properties:
        page_id:
          type: string
          format: uuid
        page_key:
          type: string
        page_name:
          type: string
        locales:
          type: array
          items:
            $ref: '#/components/schemas/LocaleLocalization'
        hreflang:
          type: object
          additionalProperties:
            type: string
        x_default:
          type:
          - string
          - 'null'
      required:
      - hreflang
      - locales
      - page_id
      - page_key
      - page_name
      - x_default
    PageSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        site_id:
          type: string
          format: uuid
        name:
          type: string
        key:
          type: string
        version:
          type: integer
        current_draft_id:
          type:
          - string
          - 'null'
          format: uuid
        current_draft_hash:
          type:
          - string
          - 'null'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
      - created_at
      - current_draft_hash
      - current_draft_id
      - id
      - key
      - name
      - site_id
      - updated_at
      - version
...
    PageTranslationList:
      type: object
      properties:
        page_id:
          type: string
          format: uuid
        default_locale:
          type: string
        supported_locales:
          type: array
          items:
            type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/PageTranslation'
      required:
      - default_locale
      - items
      - page_id
      - supported_locales
    PageTranslationSave:
      type: object
      properties:
        expected_version:
          type: integer
          minimum: 0
        slug:
          type: string
          maxLength: 80
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
        title:
          type: string
          default: ''
          maxLength: 160
        description:
          type: string
          default: ''
          maxLength: 320
        social_title:
          type: string
          default: ''
          maxLength: 160
        social_description:
          type: string
          default: ''
          maxLength: 320
        allow_title_fallback:
          type: boolean
          default: false
        allow_description_fallback:
          type: boolean
          default: false
        allow_social_title_fallback:
          type: boolean
          default: false
        allow_social_description_fallback:
          type: boolean
          default: false
      required:
      - expected_version
      - slug
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail: {}
        correlation_id:
          type:
          - string
          - 'null'
      required:
      - code
      - correlation_id
      - detail
      - status
      - title
      - type
...
    SiteCreate:
      type: object
      properties:
        name:
          type: string
          maxLength: 160
        slug:
          type: string
          maxLength: 80
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
        default_locale:
          allOf:
          - $ref: '#/components/schemas/LocaleEnum'
          default: pl
      required:
      - name
      - slug
    SiteList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SiteSummary'
        next_cursor:
          type:
          - string
          - 'null'
          format: uuid
      required:
      - items
      - next_cursor
    SiteLocalizationReport:
      type: object
      properties:
        site_id:
          type: string
          format: uuid
        default_locale:
          type: string
        supported_locales:
          type: array
          items:
            type: string
        ready_to_publish:
          type: boolean
        pages:
          type: array
          items:
            $ref: '#/components/schemas/PageLocalization'
      required:
      - default_locale
      - pages
      - ready_to_publish
      - site_id
      - supported_locales
    SiteSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        default_locale:
          type: string
        current_publication_id:
          type:
          - string
          - 'null'
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
      - created_at
      - current_publication_id
      - default_locale
      - id
      - name
      - slug
      - updated_at
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: saas_core_core_only_session